webapp/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==23.0.0
